# Issue841 remove file stream construction