Initial testing for #841. Focused specifically on the ReadByte case to figure out the pattern we want to use before trying to apply it to all of the tests.