@LoopedBard3 2021-10-01
ReadByte Default Benchmark Statistics Before & After:
Before:
|   Method | fileSize |      options |     Mean |   Error |  StdDev |   Median |      Min |      Max | Allocated |
|--------- |--------- |------------- |---------:|--------:|--------:|---------:|---------:|---------:|----------:|
| ReadByte |     1024 |         None | 147.5 us | 3.20 us | 3.55 us | 146.5 us | 143.0 us | 152.6 us |      4 KB |
| ReadByte |     1024 | Asynchronous | 153.8 us | 2.64 us | 2.71 us | 153.8 us | 149.9 us | 159.0 us |      4 KB |

After:
|   Method | fileSize |      options |     Mean |     Error |    StdDev |   Median |      Min |      Max | Allocated |
|--------- |--------- |------------- |---------:|----------:|----------:|---------:|---------:|---------:|----------:|
| ReadByte |     1024 |         None | 1.793 us | 0.0347 us | 0.0356 us | 1.780 us | 1.742 us | 1.877 us |         - |
| ReadByte |     1024 | Asynchronous | 1.785 us | 0.0220 us | 0.0184 us | 1.785 us | 1.747 us | 1.820 us |         - |

I also ran the same code but with two 16kibibyte sized files to ensure that something else wasn't causing the super low times (like accidentally only reading one byte):
|   Method | fileSize |      options |      Mean |     Error |    StdDev |    Median |       Min |       Max | Allocated |
|--------- |--------- |------------- |----------:|----------:|----------:|----------:|----------:|----------:|----------:|
| ReadByte |     1024 |         None |  1.782 us | 0.0216 us | 0.0180 us |  1.786 us |  1.749 us |  1.817 us |         - |
| ReadByte |     1024 | Asynchronous |  2.050 us | 0.0479 us | 0.0551 us |  2.039 us |  1.981 us |  2.182 us |         - |
| ReadByte |    16384 |         None | 36.135 us | 0.4380 us | 0.3657 us | 36.118 us | 35.392 us | 36.692 us |         - |
| ReadByte |    16384 | Asynchronous | 55.781 us | 0.8563 us | 0.8009 us | 55.801 us | 54.565 us | 57.342 us |     864 B |