@LoopedBard3 2021-10-02
@adamsitnik Going through some old issues in the perf repo and Bill mentioned your opinion on this change would be beneficial. What are your thoughts on using this general pattern for minimizing the creation of the FileStream during the tests? Thanks!